fix int distance

+++ app/Http/Controllers/VisitController.php
@@ -920,7 +920,7 @@ class VisitController extends Controller
             $dist = acos($dist);
             $dist = rad2deg($dist);
             $miles = $dist * 60 * 1.1515;
-            return ($miles * 1.609344)*1000;
+            return (int)($miles * 1.609344)*1000;
         }
         catch (\Exception $exception){
             return FALSE;
